renamed component to notification
