--- app/frontend/src/components/NotificationBar/_index.scss
@@ -1,1 +1,0 @@
-@forward "_NotificationBar";
