--- app/frontend/src/components/NotificationBar/index.tsx
@@ -1,1 +1,0 @@
-export { NotificationBar } from "./NotificationBar";
