+++ app/frontend/src/components/components.tsx
@@ -12,6 +12,7 @@ import { TextField } from "./Inputs/Textfield";
 import { HeaderNav } from "./Navigation/HeaderNav";
 import { FooterNav } from "./Navigation/FooterNav";
 import { ProgressBar } from "./Navigation/ProgressBar";
+import { Notification } from "./Notification/Notification";
 
 export {
   // Buttons
@@ -35,4 +36,6 @@ export {
   HeaderNav,
   FooterNav,
   ProgressBar,
+  // Notification
+  Notification,
 };
