+++ app/frontend/src/index.scss
@@ -11,7 +11,7 @@
 @use "components/Dialog/Dialog";
 @use "components/Inputs" as *;
 @use "components/Navigation" as *;
-@use "components/NotificationBar" as *;
+@use "components/Notification/Notification";
 
 // Import page styles
 @use "pages/Demo/Demo";
